requirements.txt
numpy>=1.24
pydantic>=2.6
scikit-learn>=1.3
PyYAML>=6.0
pytest>=7.0
# Optional LLM provider; install only if you plan to call OpenAI:
# openai>=1.30
